requirements.txt
appdirs==1.4.0
carmen==0.0.4
coreapi==2.2.4
coreschema==0.0.4
Django==1.10.5
django-bootstrap3==8.2.1
django-filter==1.0.1
django-jsonview==0.4.3
django-rest-swagger==2.1.1
-e git+https://github.com/michaelbrooks/django-twitter-stream.git@c0285e1663f26a3e845a11131a4ebe4ec9e9d9e6#egg=django_twitter_stream
djangorestframework==3.5.4
funcsigs==1.0.2
future==0.16.0
geopy==1.11.0
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==0.23
mock==2.0.0
nltk==3.2.2
oauthlib==2.0.1
openapi-codec==1.3.1
packaging==16.8
pbr==1.10.0
py==1.4.32
pycryptodomex==3.4.3
pyjwkest==1.3.2
PyJWT==1.4.2
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
swapper==0.1.1
textblob==0.12.0
twarc==1.0.7
tweepy==3.5.0
twitter-monitor==0.3.0
unicodecsv==0.14.1
uritemplate==3.0.0
whitenoise==3.3.0
